+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0
 
 * Make language-related functionality "multi-lingual", i.e. a serialization chunk can contain multiple `Language`s.
+  * (Languages will be sorted by name.)
 * Improve extraction functionality: only catch JSON-parsing exceptions.
 * Configure single entrypoint named `lionweb-cli`.